+++ config/nvim/init.lua
@@ -1,20 +1,6 @@
-require('basic')
-require('options')
-require('keymaps')
+require('config.basic')
+require('config.options')
+require('config.keymaps')
 
-local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
-if not vim.loop.fs_stat(lazypath) then
-  vim.fn.system({
-    "git",
-    "clone",
-    "--filter=blob:none",
-    "https://github.com/folke/lazy.nvim.git",
-    "--branch=stable", -- latest stable release
-    lazypath,
-  })
-end
-vim.opt.rtp:prepend(lazypath)
-
-plugins = require('plugins/core')
-
-require('lazy').setup(plugins)
+-- Plugins
+require("config.lazy")
